refactor(theaters): unify the domain and db entity in a single class
