--- src/modules/theaters/entities/theater.ts
@@ -1,6 +1,0 @@
-export class Theater {
-  constructor(
-    public readonly id: string,
-    public readonly name: string,
-  ) {}
-}
