+++ src/modules/theaters/repositories/interfaces/theather.repository.ts
@@ -1,4 +1,4 @@
-import { Theater } from '../../entities/theater.ts';
+import { Theater } from '../../database/theater.entity.ts';
 
 export interface ITheaterRepository {
   findById(id: string): Promise<Theater | null>;
