+++ src/modules/theaters/repositories/theater.repository.ts
@@ -1,13 +1,13 @@
-import { TheaterDBEntity } from '../database/theater.entity.ts';
-import { Theater } from '../entities/theater.ts';
+import { appDataSource } from '../../shared/data-source/data-source.ts';
+import { Theater } from '../database/theater.entity.ts';
 import { ITheaterRepository } from './interfaces/theather.repository.ts';
 
 export class TheaterRepository implements ITheaterRepository {
   async findById(id: string): Promise<Theater | null> {
-    const theater = await TheaterDBEntity.findOne({ where: { id } });
+    const theater = await appDataSource.getRepository(Theater).findOne({ where: { id } });
 
     if (!theater) return null;
 
-    return new Theater(theater.id, theater.name);
+    return theater;
   }
 }
